Visual Primacy Bias Code/src/baselines/direct.py
"""Direct Instruction baseline attack.

This is the simplest baseline: directly asking the model to perform harmful tasks
without any jailbreaking techniques.
"""

from typing import Optional, Tuple
from dataclasses import dataclass
from PIL import Image
import numpy as np


@dataclass
class DirectInstructionConfig:
    """Configuration for Direct Instruction baseline."""
    image_size: Tuple[int, int] = (512, 512)
    use_blank_image: bool = True


class DirectInstructionAttack:
    """Direct Instruction baseline attack.
    
    This baseline simply presents the harmful instruction directly to the model
    without any obfuscation or jailbreaking techniques. It serves as a lower
    bound for attack success rate comparison.
    """
    
    def __init__(self, config: Optional[DirectInstructionConfig] = None):
        self.config = config or DirectInstructionConfig()
    
    def _create_blank_image(self) -> Image.Image:
        """Create a blank white image."""
        return Image.new('RGB', self.config.image_size, (255, 255, 255))
    
    def _create_random_image(self) -> Image.Image:
        """Create a random noise image."""
        noise = np.random.randint(0, 256, (*self.config.image_size, 3), dtype=np.uint8)
        return Image.fromarray(noise)
    
    def attack(
        self,
        harmful_instruction: str,
        image: Optional[Image.Image] = None
    ) -> Tuple[Image.Image, str]:
        """Generate Direct Instruction attack.
        
        Args:
            harmful_instruction: The harmful instruction to present directly
            image: Optional image to use (if None, creates blank/random image)
            
        Returns:
            Tuple of (image, prompt)
        """
        if image is None:
            if self.config.use_blank_image:
                image = self._create_blank_image()
            else:
                image = self._create_random_image()
        
        # Direct instruction without any jailbreaking
        prompt = harmful_instruction
        
        return image, prompt
    
    def attack_with_context(
        self,
        harmful_instruction: str,
        context: str = "Please help me with the following:",
        image: Optional[Image.Image] = None
    ) -> Tuple[Image.Image, str]:
        """Generate Direct Instruction attack with benign context.
        
        Args:
            harmful_instruction: The harmful instruction
            context: Benign context prefix
            image: Optional image to use
            
        Returns:
            Tuple of (image, prompt)
        """
        if image is None:
            image = self._create_blank_image()
        
        prompt = f"{context}\n\n{harmful_instruction}"
        
        return image, prompt
    
    def attack_variants(
        self,
        harmful_instruction: str
    ) -> list:
        """Generate multiple variants of direct instruction attacks.
        
        Args:
            harmful_instruction: The harmful instruction
            
        Returns:
            List of (image, prompt) tuples
        """
        variants = []
        
        # Variant 1: Plain instruction
        variants.append((
            self._create_blank_image(),
            harmful_instruction
        ))
        
        # Variant 2: With polite prefix
        variants.append((
            self._create_blank_image(),
            f"Please {harmful_instruction.lower()}"
        ))
        
        # Variant 3: As a question
        variants.append((
            self._create_blank_image(),
            f"How do I {harmful_instruction.lower()}?"
        ))
        
        # Variant 4: With random image
        variants.append((
            self._create_random_image(),
            harmful_instruction
        ))
        
        return variants
